Take in Readme

+++ organise.py
@@ -38,5 +38,40 @@ def get_new_filename(f, create_dirs=False):
 for f in files:
     if re_f.search(f):
         f_ = get_new_filename(f, create_dirs=True)
-        print(f"{f:70} > {f_}")
-        os.rename(f, f_)
+        if not os.path.exists(f_):
+            print(f"{f:70} > {f_}")
+            os.rename(f, f_)
+        else:
+            print(f"{f:70} > {f_}: Exception: path exists")
+
+header = [
+        "# Leet Code Solutions",
+        "I started doing leet code properly on *03.10.2021.*",
+        "Here I will document all the problems that I have solved using *jupyter-notebooks*."
+    ]
+
+body = []
+
+import re
+count = 0
+for r, dirs, files in os.walk("./"):
+    if not r.startswith("./.") and r != "./":
+        line = f"## {r} \n"
+        body.append(line)
+        line = f"*{len(files)} problems* \n"
+        body.append(line)
+        count += len(files)
+        for f in files:
+            v = re.split(r"_|\.", f)
+            v_ = " ".join(v[1:-1])
+            line = f"* {v[0]} [{v_}]({r}/{f}) \n"
+            body.append(line)
+
+
+with open("Readme.md", "w+") as fh:
+    for l in header:
+        fh.write(l + "\n")
+    line = f"* A total of **{count} problems** \n"
+    fh.write(line + "")
+    for l in body:
+        fh.write(l + "")
